net/mlx5e: Unregister all vport reps in one call
We can just do this in one call.

Also, since the unreg flow involves unloading, make sure to do it in
decreasing order such that the uplink rep (currently the PF) is the last
one unregistered. This goes along the lines of commit c191220396db8
("net/mlx5: E-Switch, Unload the representors in the correct order").

Issue: 1306741
Change-Id: I36a6f6d6f4b5105968ab29aaf2d94ac06ac45b13
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>