net/mlx5: Modify page management logic
This patch supports the option for a PF to get its ICM pages from
another pci function. The other pci function is called ECPF (Embedded
CPU Function) and the driver can identify if it is running on such a
function by reading a bit in the initialization segment.

When firmware asks for pages it would issue a page request event
specifying how many pages it requests and for which function. That
driver responds with a manage_pages command providing the requested
pages along with an indication for which function it is providing these
pages.

The encoding before this patch was as follows:
function_id == 0: pages are requested for the function receiving the
EQE.
function_id != 0: pages are requested for VF identified by the
function_id value

A new one bit field in the EQE identifies that pages are requested for
the ECPF.

The notion of page_supplier can be introduced here and to support that,
manage pages and query pages were modified so firmware can distinguish
the following cases:

1. Function provides pages for itself
2. PF provides pages for its VF
3. ECPF provides pages to itself
4. ECPF provides pages for another function

This distinction is possible through the introduction of the bit
"embedded_cpu_function" in query_pages, manage_pages and page request
EQE.

Issue: 1272611
Change-Id: Id5776d3f54e48e6d3d5d4e0ea0f94a7337daa5ff
Signed-off-by: Eli Cohen <eli@mellanox.com>

Conflicts:
	include/linux/mlx5/driver.h