net/mlx5: Update enable_hca dependency
With the introduction of ECPF, where the ECPF can be the page supplier
for a PF, we require that the ECPF driver will call enable/disable HCA
for that PF in the same way a PF does this for its VFs. The PF is still
responsible for calling enable/disable HCA for its VFs.

To distinguish between the ECPF executing enable/disable HCA for itself
or for the PF, it sets the embedded_cpu_function bit in the input params
struct of these commands. When the bit is cleared and func_id is zero,
it refers to the peer PF.

Issue: 1272611
Change-Id: Ie21dd051011bbd53c07fb73ed1a7514e00c93af9
Signed-off-by: Eli Cohen <eli@mellanox.com>