net/mlx5: Implement query host parameters
In Embedded CPU (EC) configurations, the EC driver needs to know when
the number of virtual functions change on the corresponding PF at the
host side. This is required so the EC driver can create or destroy
representor net devices that represent the VFs ports.

Whenever a change in the number of VFs occurs, firmware will generate an
event towards the EC which will trigger a work to complete the rest of
the handling. The specifics of the handling will be introduced in a
downstream patch.

Issue: 1272611
Change-Id: I326be31a940ffd4624689212366c8c77b409a65e
Signed-off-by: Eli Cohen <eli@mellanox.com>