net/mlx5: E-Switch, Re-arrange representors positions
The current implementation puts the uplink representor at index zero of the
vport_reps array. It is not "natural" to place it at index 0 since we
want to put the representor for vport0 at index 0; this can happen with
the introduction of SmartNIC but also if future implementation will set
a VF other than VF0 as the eswitch manager.

The change invloves defining the macro MLX5_TOTAL_REPS to evaluate the
number of representors required and decouple it from the number of
vports.

Also, both MLX5_TOTAL_REPS and MLX5_TOTAL_VPORTS are modified to
provide place for vport 0.

Issue: 1272611
Change-Id: I304620c7126a9c40ddfa9cacd7675b13268ca00d
Signed-off-by: Eli Cohen <eli@mellanox.com>