IB/mlx5: Modify ib_rep to support smartnic model
Modify representors states according to the new scheme. With the new
scheme, a representor can be in on of the following states:
- unregistered
- registered
- enabled
- loaded

Now, that we may have a representor for vport 0, we extend the range of
registered vports to include vport 0 and also ecpf for the case where
the eswitch manager is running on the host.

We also change the prototype of mlx5_ib_get_rep_ibdev() and
mlx5_ib_get_rep_netdev() to accept vport number instead of vport index.
This matches mlx5_eswitch_get_proto_dev() which is used to return the
actual pointer to the device.

Lastly, we add sysfs entries to mlx5 IB devices to show the representor
state and the vport this representor is associated with.

Issue: 1272611
Change-Id: I5d267f18f820652f8903ad90b54b040d48a71b60
Signed-off-by: Eli Cohen <eli@mellanox.com>