net/mlx5: E-Switch, Allow adding UC addresses for any vport
In SmartNIC mode, the eswitch manager is not necessarily the PF (vport
0). To allow it to add/del unicast addresses for any vport, we add a
helper function that returns a boolean indicating if the vport is the
eswitch manager and use it to skip updating its own UC MACs.

ECPF stands for Embedded CPU Physical Function. It is a physical
function that connects to the eswitch through a port having the value
0xfffe. The ECPF can assigned the eswitch manager through NV config
variable ECPF_ESWITCH_MANAGER.

mlx5_ecpf_vport_exists() helper function is added to the case where the
eswitch manager is the PF (e.g. the pci function connected to vport 0);
the PF needs to know on the existence of this port since it controls it.

Issue: 1272611
Change-Id: I7dc68c51e99164a56f4abf3bc30ea098a385250e
Signed-off-by: Eli Cohen <eli@mellanox.com>