net/mlx5: E-Switch, Re-read device caps after FDB create/destroy
Re-read device capabilities after eswitch tables are created or
destroyed to reflect capability changes that the firmware may apply, e.g
due to tables created with encap flag set. It required for ECPF since it
creates the FDB before the NIC netdevice is initialized. This change
will allow it to read the correct capabilities.

Issue: 1431255
Change-Id: I23e4f63fc6e737ec106127873ad96bee500c120f
Signed-off-by: Eli Cohen <eli@mellanox.com>