net/mlx5: E-Switcn, Fix reference to my vport number
Functions referring to vport numbers used to pass 0 as the vport number
argument to denote that they refer to their own vport numbers. So far it
was sufficient but now with ECPF being able to access port 0, this is not
enough anymore. Use real vport number and let the callee encode the
correct values in the command so it refers to the right vport number.

Issue: 1272611
Change-Id: I2fdd696080ab536ff5cd2c535fc255c0800a160c
Signed-off-by: Eli Cohen <eli@mellanox.com>

Conflicts:
	drivers/net/ethernet/mellanox/mlx5/core/en_main.c
	drivers/net/ethernet/mellanox/mlx5/core/en_selftest.c