net/mlx5: E-Switch, Modify vport encoding in commands
Commands referring to vports use the following scheme:

1. When referring to my own vport, put 0 in vport and 0 in other_vport.
2. When referring to another vport, put the vport number of the referred
   vport and put 1 in other_vport.

Some functions made use of this knowledge when they wanted to access
their own vport by passing 0. This is not good since it does not allow
to distinguish between the case when ECPF wants to access vport 0 and
the case where a function wants to access its own vport.

This get_other_vport() is defined to return the correct setting for the
other_vport fields and modify the vport_num value to correctly encode
the desired vport number.

Issue: 1272611
Change-Id: I7280647cf23510ba43e46f6524e8c907c90234bb
Signed-off-by: Eli Cohen <eli@mellanox.com>